+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "smartypay-client-react",
-  "version": "1.0.10",
+  "version": "1.0.11",
   "description": "React library for show a custom payment button",
   "author": "Evgeny Dolganov <evgenij.dolganov@gmail.com>",
   "repository": "https://github.com/smarty-pay/smartypay-client-react",
@@ -42,6 +42,6 @@
   "dependencies": {
     "react": ">=14.0.0",
     "react-dom": ">=14.0.0",
-    "smartypay-client-sdk": ">=1.0.4"
+    "smartypay-client-sdk": ">=1.0.12"
   }
 }
